moving first chapter parter no relevant class
